# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FACTA — Independent product intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit an unknown product for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get submission status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/confirm-ocr:
    patch:
      operationId: confirmOcr
      tags: [submissions]
      summary: User confirms or corrects OCR results
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrConfirmation"
      responses:
        "200":
          description: OCR confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/finalize:
    post:
      operationId: finalizeSubmission
      tags: [submissions]
      summary: Instantly create a provisional product from a confirmed submission and score it
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Provisional product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizeResult"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /submissions/ocr:
    post:
      operationId: processOcr
      tags: [submissions]
      summary: Run OCR on uploaded image and return extracted text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrRequest"
      responses:
        "200":
          description: OCR result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrResult"

  # ── User Preferences ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    FinalizeResult:
      type: object
      required: [productId]
      properties:
        productId: { type: integer }
        overallScore: { type: integer }
        scoreGrade: { type: string }
...
    OcrRequest:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:  { type: string }
        imageType:    { type: string, default: "ingredients" }
    OcrResult:
      type: object
      required: [extractedText, confidence]
      properties:
        extractedText:      { type: string }
        confidence:         { type: number }
        structuredData:
          type: object
          additionalProperties: true
        rawIngredients:     { type: ["string", "null"] }
        parsedNutrition:
          type: object
          additionalProperties: true
    SubmissionInput:
      type: object
      required: [productName]
      properties:
        productName:      { type: string }
        brandName:        { type: ["string", "null"] }
        barcode:          { type: ["string", "null"] }
        retailerSlug:     { type: ["string", "null"] }
        frontImageBase64: { type: ["string", "null"] }
        ingredientsImageBase64: { type: ["string", "null"] }
        nutritionImageBase64:   { type: ["string", "null"] }
        rawIngredientsText: { type: ["string", "null"] }
        userSession:      { type: ["string", "null"] }
        userConsented:    { type: boolean, default: false }
    OcrConfirmation:
      type: object
      required: [confirmedIngredients]
      properties:
        confirmedIngredients: { type: string }
        confirmedNutrition:
          type: object
          additionalProperties: true
        userCorrections:  { type: ["string", "null"] }
    Submission:
      type: object
      required: [id, productName, status, createdAt]
      properties:
        id:              { type: integer }
        productName:     { type: string }
        brandName:       { type: ["string", "null"] }
        barcode:         { type: ["string", "null"] }
        status:          { type: string }
        ocrStatus:       { type: ["string", "null"] }
        extractedIngredients: { type: ["string", "null"] }
        provisionalScore: { type: ["integer", "null"] }
        provisionalGrade: { type: ["string", "null"] }
        dataCompleteness: { type: ["number", "null"] }
        userSession:     { type: ["string", "null"] }
        reviewNote:      { type: ["string", "null"] }
        createdAt:       { type: string }
        updatedAt:       { type: string }

    # ── User Preferences ──────────────────────────────────────────────────
...